refactor new authority model, convert some components, fix bugs
