--- packages/engine/src/networking/components/NetworkObjectAuthorityTag.ts
@@ -1,3 +1,0 @@
-import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
-
-export const NetworkObjectAuthorityTag = createMappedComponent<true>('NetworkObjectAuthorityTag')
