--- packages/engine/src/networking/components/NetworkObjectOwnedTag.ts
@@ -1,3 +1,0 @@
-import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
-
-export const NetworkObjectOwnedTag = createMappedComponent<true>('NetworkObjectOwnedTag')
