+++ packages/engine/src/networking/functions/WorldNetworkActionReceptor.test.ts
@@ -13,7 +13,7 @@ import { createEngine } from '../../initializeEngine'
 import { Physics } from '../../physics/classes/Physics'
 import { NetworkTopics } from '../classes/Network'
 import { NetworkObjectComponent } from '../components/NetworkObjectComponent'
-import { NetworkObjectOwnedTag } from '../components/NetworkObjectOwnedTag'
+import { NetworkObjectOwnedTag } from '../components/NetworkObjectComponent'
 import WorldNetworkActionSystem from '../systems/WorldNetworkActionSystem'
 import { NetworkPeerFunctions } from './NetworkPeerFunctions'
 import { WorldNetworkAction } from './WorldNetworkAction'
